+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules.hs
@@ -1,6 +1,9 @@
 module Cardano.Ledger.Conway.Rules (
   module Cardano.Ledger.Conway.Rules.Cert,
-  module Cardano.Ledger.Conway.Rules.Delegs,
+  module Cardano.Ledger.Conway.Rules.Deleg,
+  module Cardano.Ledger.Conway.Rules.Pool,
+  module Cardano.Ledger.Conway.Rules.VDel,
+  module Cardano.Ledger.Conway.Rules.Certs,
   module Cardano.Ledger.Conway.Rules.Enact,
   module Cardano.Ledger.Conway.Rules.Epoch,
   module Cardano.Ledger.Conway.Rules.Ledger,
@@ -13,12 +16,15 @@ module Cardano.Ledger.Conway.Rules (
 where
 
 import Cardano.Ledger.Conway.Rules.Cert
-import Cardano.Ledger.Conway.Rules.Delegs
+import Cardano.Ledger.Conway.Rules.Certs
+import Cardano.Ledger.Conway.Rules.Deleg
 import Cardano.Ledger.Conway.Rules.Enact
 import Cardano.Ledger.Conway.Rules.Epoch
 import Cardano.Ledger.Conway.Rules.Ledger
 import Cardano.Ledger.Conway.Rules.NewEpoch
+import Cardano.Ledger.Conway.Rules.Pool
 import Cardano.Ledger.Conway.Rules.Ratify
 import Cardano.Ledger.Conway.Rules.Tally
 import Cardano.Ledger.Conway.Rules.Tickf
 import Cardano.Ledger.Conway.Rules.Utxos
+import Cardano.Ledger.Conway.Rules.VDel